Rename `ProjectTiles` to `ProjectCards`
