+++ web-common/src/features/welcome/ProjectCards.svelte
